backend/requirements.txt
Flask
numpy
Pillow
python-multipart
python-dotenv
openai
requests
flask-swagger-ui
flask-cors
pymongo==4.6.0